Create new component SearchResultAttachmentsList which is responsible for the UI of attachment list.
Also move the support of viewing next/previous attachments in lightbox
to this component.
