+++ Source/Plugins/Core/com.equella.core/js/tsrc/search/components/SearchResult.tsx
@@ -16,39 +16,27 @@
  * limitations under the License.
  */
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Badge,
   Divider,
   Grid,
   Hidden,
   IconButton,
   List,
   ListItem,
-  ListItemIcon,
-  ListItemSecondaryAction,
   ListItemText,
   Theme,
   Typography,
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import Tooltip from "@material-ui/core/Tooltip";
-import AttachFile from "@material-ui/icons/AttachFile";
 import DragIndicatorIcon from "@material-ui/icons/DragIndicator";
-import ExpandMore from "@material-ui/icons/ExpandMore";
 import FavoriteIcon from "@material-ui/icons/Favorite";
 import FavoriteBorderIcon from "@material-ui/icons/FavoriteBorder";
-import InsertDriveFile from "@material-ui/icons/InsertDriveFile";
-import Search from "@material-ui/icons/Search";
 import * as OEQ from "@openequella/rest-api-client";
 import * as React from "react";
-import { SyntheticEvent, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import ReactHtmlParser from "react-html-parser";
 import { HashLink } from "react-router-hash-link";
 import { sprintf } from "sprintf-js";
 import { Date as DateDisplay } from "../../components/Date";
-import ItemAttachmentLink from "../../components/ItemAttachmentLink";
 import { OeqLink } from "../../components/OeqLink";
 import OEQThumb from "../../components/OEQThumb";
 import { StarRating } from "../../components/StarRating";
@@ -60,9 +48,7 @@ import {
 } from "../../modules/FavouriteModule";
 import {
   buildSelectionSessionItemSummaryLink,
-  getSearchPageAttachmentClass,
   getSearchPageItemClass,
-  isSelectionSessionInSkinny,
   isSelectionSessionInStructured,
   isSelectionSessionOpen,
   isSelectSummaryButtonDisabled,
@@ -78,11 +64,7 @@ import type {
   FavouriteItemInfo,
 } from "./FavouriteItemDialog";
 import { ResourceSelector } from "./ResourceSelector";
-import {
-  determineAttachmentViewUrl,
-  determineViewer,
-  ViewerDefinition,
-} from "../../modules/ViewerModule";
+import { SearchResultAttachmentsList } from "./SearchResultAttachmentsList";
 
 const {
   searchResult: searchResultStrings,
@@ -112,21 +94,6 @@ const useStyles = makeStyles((theme: Theme) => {
     },
     status: {
       textTransform: "capitalize",
-    },
-    nested: {
-      paddingLeft: theme.spacing(4),
-    },
-    attachmentExpander: {
-      marginTop: theme.spacing(2),
-      marginBottom: theme.spacing(2),
-    },
-    attachmentBadge: {
-      backgroundColor: theme.palette.background.paper,
-      color: theme.palette.secondary.main,
-      borderRadius: "50%",
-    },
-    attachmentListItem: {
-      width: "100%",
     },
     highlight: {
       color: theme.palette.secondary.main,
@@ -162,22 +129,14 @@ export interface SearchResultProps {
   onFavouriteItem: (itemInfo: FavouriteItemInfo) => void;
 }
 
-export interface AttachmentAndViewer {
-  /**
-   * The details of an attachment.
-   */
-  attachment: OEQ.Search.Attachment;
-  /**
-   * Viewer information for the attachment, including which viewer to be used and the view URL.
-   */
-  viewer: ViewerDefinition;
-}
-
 export default function SearchResult({
   getViewerDetails = getMimeTypeDefaultViewerDetails,
   handleError,
   highlights,
-  item: {
+  item,
+  onFavouriteItem,
+}: SearchResultProps) {
+  const {
     name,
     version,
     uuid,
@@ -187,111 +146,28 @@ export default function SearchResult({
     status,
     displayOptions,
     attachments = [],
-    keywordFoundInAttachment,
     commentCount = 0,
     starRatings,
     bookmarkId: bookmarkDefaultId,
     isLatestVersion,
-  },
-  onFavouriteItem,
-}: SearchResultProps) {
+  } = item;
   const itemKey = `${uuid}/${version}`;
   const classes = useStyles();
   const inSelectionSession: boolean = isSelectionSessionOpen();
   const inStructured = isSelectionSessionInStructured();
-  const inSkinny = isSelectionSessionInSkinny();
-
-  const [attachExpanded, setAttachExpanded] = useState(
-    (inSelectionSession
-      ? displayOptions?.integrationOpen
-      : displayOptions?.standardOpen) ?? false
-  );
-  const [attachmentsAndViewers, setAttachmentsAndViewers] = useState<
-    AttachmentAndViewer[]
-  >([]);
 
   const [bookmarkId, setBookmarkId] = useState<number | undefined>(
     bookmarkDefaultId
   );
 
-  // Responsible for determining the MIME type viewer for the provided attachments
-  useEffect(() => {
-    let mounted = true;
-
-    if (!attachments.length) {
-      // If there are no attachments, skip this effect
-      return;
-    }
-
-    const transform = async (
-      a: OEQ.Search.Attachment
-    ): Promise<AttachmentAndViewer> => {
-      const {
-        attachmentType,
-        mimeType,
-        links: { view: defaultViewUrl },
-        filePath,
-      } = a;
-      const viewUrl = determineAttachmentViewUrl(
-        uuid,
-        version,
-        attachmentType,
-        defaultViewUrl,
-        filePath
-      );
-
-      let viewerDetails: OEQ.MimeType.MimeTypeViewerDetail | undefined;
-      try {
-        viewerDetails = a.mimeType
-          ? await getViewerDetails(a.mimeType)
-          : undefined;
-      } catch (error) {
-        handleError({
-          ...error,
-          message: `${searchResultStrings.errors.getAttachmentViewerDetailsFailure}: ${error.message}`,
-        });
-      }
-
-      return {
-        attachment: a,
-        viewer: determineViewer(
-          attachmentType,
-          viewUrl,
-          mimeType,
-          viewerDetails?.viewerId
-        ),
-      };
-    };
-
-    (async () => {
-      const attachmentsAndViewers = await Promise.all(
-        attachments.map<Promise<AttachmentAndViewer>>(transform)
-      );
-      if (mounted) {
-        setAttachmentsAndViewers(attachmentsAndViewers);
-      }
-    })();
-
-    return () => {
-      // Short circuit if this component is unmounted before all its comms are done.
-      mounted = false;
-    };
-  }, [attachments, getViewerDetails, handleError, uuid, version]);
-
   // In Selection Session, make each attachment draggable.
   useEffect(() => {
     if (inStructured) {
-      attachmentsAndViewers.forEach(({ attachment }) => {
+      attachments.forEach((attachment) => {
         prepareDraggable(attachment.id, false);
       });
     }
-  }, [attachmentsAndViewers, inStructured]);
-
-  const handleAttachmentPanelClick = (event: SyntheticEvent) => {
-    /** prevents the SearchResult onClick from firing when attachment panel is clicked */
-    event.stopPropagation();
-    setAttachExpanded(!attachExpanded);
-  };
+  }, [attachments, inStructured]);
 
   const handleSelectResource = (
     itemKey: string,
@@ -416,120 +292,6 @@ export default function SearchResult({
     }
   );
 
-  const generateAttachmentList = () => {
-    const attachmentsList = attachmentsAndViewers.map(
-      (attachmentAndViewer: AttachmentAndViewer) => {
-        const {
-          attachment: { id, description },
-        } = attachmentAndViewer;
-        return (
-          <ListItem
-            key={id}
-            id={id}
-            button
-            className={`${classes.nested} ${getSearchPageAttachmentClass()}`} // Give a class so each attachment can be dropped to the course list.
-            data-itemuuid={uuid} // These 'data-xx' attributes are used in the 'dropCallBack' of 'courselist.js'.
-            data-itemversion={version}
-            data-attachmentuuid={id}
-          >
-            <ListItemIcon>
-              {inStructured ? <DragIndicatorIcon /> : <InsertDriveFile />}
-            </ListItemIcon>
-            <ItemAttachmentLink
-              selectedAttachment={attachmentAndViewer}
-              allLightBoxAttachments={attachmentsAndViewers.filter(
-                (av) => av.viewer[0] === "lightbox"
-              )}
-            >
-              <ListItemText color="primary" primary={description} />
-            </ItemAttachmentLink>
-            {inSelectionSession && (
-              <ListItemSecondaryAction>
-                <ResourceSelector
-                  labelText={selectResourceStrings.attachment}
-                  isStopPropagation
-                  onClick={() => {
-                    handleSelectResource(itemKey, [id]);
-                  }}
-                />
-              </ListItemSecondaryAction>
-            )}
-          </ListItem>
-        );
-      }
-    );
-
-    const accordionText = (
-      <Typography>{searchResultStrings.attachments}</Typography>
-    );
-
-    const accordionSummaryContent = inSelectionSession ? (
-      <Grid container alignItems="center">
-        <Grid item>{accordionText}</Grid>
-        <Grid>
-          {!inSkinny && (
-            <ResourceSelector
-              labelText={selectResourceStrings.allAttachments}
-              isStopPropagation
-              onClick={() => {
-                const attachments = attachmentsAndViewers.map(
-                  ({ attachment }) => attachment.id
-                );
-                handleSelectResource(itemKey, attachments);
-              }}
-            />
-          )}
-        </Grid>
-      </Grid>
-    ) : (
-      accordionText
-    );
-
-    const attachFileBadge = (includeIndicator: boolean) => (
-      <Badge
-        anchorOrigin={{
-          vertical: "bottom",
-          horizontal: "right",
-        }}
-        overlap="circle"
-        badgeContent={
-          includeIndicator ? (
-            <Tooltip
-              title={searchResultStrings.keywordFoundInAttachment}
-              aria-label={searchResultStrings.keywordFoundInAttachment}
-            >
-              <Search fontSize="small" className={classes.attachmentBadge} />
-            </Tooltip>
-          ) : undefined
-        }
-      >
-        <AttachFile />
-      </Badge>
-    );
-
-    if (attachmentsList.length > 0)
-      return (
-        <Accordion
-          className={classes.attachmentExpander}
-          expanded={attachExpanded}
-          onClick={(event) => handleAttachmentPanelClick(event)}
-        >
-          <AccordionSummary expandIcon={<ExpandMore />}>
-            <Grid container spacing={2} alignItems="center">
-              <Grid item>{attachFileBadge(keywordFoundInAttachment)}</Grid>
-              <Grid item>{accordionSummaryContent}</Grid>
-            </Grid>
-          </AccordionSummary>
-          <AccordionDetails>
-            <List disablePadding className={classes.attachmentListItem}>
-              {attachmentsList}
-            </List>
-          </AccordionDetails>
-        </Accordion>
-      );
-    return null;
-  };
-
   const highlightField = (fieldValue: string) =>
     ReactHtmlParser(highlight(fieldValue, highlights, classes.highlight));
 
@@ -602,7 +364,11 @@ export default function SearchResult({
               {highlightField(description ?? "")}
             </Typography>
             <List disablePadding>{customDisplayMetadata}</List>
-            {generateAttachmentList()}
+            <SearchResultAttachmentsList
+              item={item}
+              handleError={handleError}
+              getViewerDetails={getViewerDetails}
+            />
             {generateItemMetadata()}
           </>
         }
